+++ src/FluxFlow.Components.Timers.Composition/TimersCompositionNodeRegistryExtensions.cs
@@ -1,7 +1,7 @@
-using FluxFlow.Components.Timers.Contracts;
 using FluxFlow.Components.Timers.Nodes;
 using FluxFlow.Components.Timers.Options;
 using FluxFlow.Composition;
+using FluxFlow.Data;
 
 namespace FluxFlow.Components.Timers.Composition;
 
@@ -19,7 +19,7 @@ public static class TimersCompositionNodeRegistryExtensions
             CreateTimerIntervalNode,
             outputs:
             [
-                CompositionPorts.Metadata<TimerTick>(
+                CompositionPorts.Metadata<FlowValue>(
                     TimersCompositionPortNames.Output)
             ]);
     }
@@ -36,7 +36,7 @@ public static class TimersCompositionNodeRegistryExtensions
             CreateTimerScheduleNode,
             outputs:
             [
-                CompositionPorts.Metadata<ScheduleTick>(
+                CompositionPorts.Metadata<FlowValue>(
                     TimersCompositionPortNames.Output)
             ]);
     }
@@ -63,6 +63,28 @@ public static class TimersCompositionNodeRegistryExtensions
             ]);
     }
 
+    public static CompositionNodeRegistry RegisterTimerDelay(
+        this CompositionNodeRegistry registry,
+        string nodeType = TimersCompositionNodeTypes.Delay)
+    {
+        ArgumentNullException.ThrowIfNull(registry);
+        ArgumentException.ThrowIfNullOrWhiteSpace(nodeType);
+
+        return registry.Register(
+            nodeType,
+            CreateFlowValueTimerDelayNode,
+            inputs:
+            [
+                CompositionPorts.Metadata<FlowValue>(
+                    TimersCompositionPortNames.Input)
+            ],
+            outputs:
+            [
+                CompositionPorts.Metadata<FlowResult<FlowValue>>(
+                    TimersCompositionPortNames.Output)
+            ]);
+    }
+
     public static CompositionNodeRegistry RegisterTimerThrottle<TInput>(
         this CompositionNodeRegistry registry,
         string nodeType = TimersCompositionNodeTypes.Throttle)
@@ -81,6 +103,28 @@ public static class TimersCompositionNodeRegistryExtensions
             outputs:
             [
                 CompositionPorts.Metadata<TInput>(
+                    TimersCompositionPortNames.Output)
+            ]);
+    }
+
+    public static CompositionNodeRegistry RegisterTimerThrottle(
+        this CompositionNodeRegistry registry,
+        string nodeType = TimersCompositionNodeTypes.Throttle)
+    {
+        ArgumentNullException.ThrowIfNull(registry);
+        ArgumentException.ThrowIfNullOrWhiteSpace(nodeType);
+
+        return registry.Register(
+            nodeType,
+            CreateFlowValueTimerThrottleNode,
+            inputs:
+            [
+                CompositionPorts.Metadata<FlowValue>(
+                    TimersCompositionPortNames.Input)
+            ],
+            outputs:
+            [
+                CompositionPorts.Metadata<FlowResult<FlowValue>>(
                     TimersCompositionPortNames.Output)
             ]);
     }
@@ -107,24 +151,45 @@ public static class TimersCompositionNodeRegistryExtensions
             ]);
     }
 
+    public static CompositionNodeRegistry RegisterTimerDebounce(
+        this CompositionNodeRegistry registry,
+        string nodeType = TimersCompositionNodeTypes.Debounce)
+    {
+        ArgumentNullException.ThrowIfNull(registry);
+        ArgumentException.ThrowIfNullOrWhiteSpace(nodeType);
+
+        return registry.Register(
+            nodeType,
+            CreateFlowValueTimerDebounceNode,
+            inputs:
+            [
+                CompositionPorts.Metadata<FlowValue>(
+                    TimersCompositionPortNames.Input)
+            ],
+            outputs:
+            [
+                CompositionPorts.Metadata<FlowResult<FlowValue>>(
+                    TimersCompositionPortNames.Output)
+            ]);
+    }
+
     private static ValueTask<ComposedNode> CreateTimerIntervalNode(
         CompositionNodeFactoryContext context)
     {
         var settings = context.BindConfiguration<TimerIntervalSettings>();
         var clock = context.GetResource<TimeProvider>(
             TimersCompositionResourceNames.Clock);
-        var node = new TimerIntervalNode(settings, clock);
+        var node = new FlowValueTimerIntervalNode(settings, clock);
 
         return ValueTask.FromResult(ComposedNode.Create(
             node,
             outputs:
             [
-                CompositionPorts.Output<TimerTick>(
+                CompositionPorts.Output<FlowValue>(
                     TimersCompositionPortNames.Output,
                     node.Output)
             ],
-            events: node.Events,
-            errors: node.Errors));
+            events: node.Events));
     }
 
     private static ValueTask<ComposedNode> CreateTimerScheduleNode(
@@ -133,18 +198,92 @@ public static class TimersCompositionNodeRegistryExtensions
         var settings = context.BindConfiguration<TimerScheduleSettings>();
         var clock = context.GetResource<TimeProvider>(
             TimersCompositionResourceNames.Clock);
-        var node = new TimerScheduleNode(settings, clock);
+        var node = new FlowValueTimerScheduleNode(settings, clock);
 
         return ValueTask.FromResult(ComposedNode.Create(
             node,
             outputs:
             [
-                CompositionPorts.Output<ScheduleTick>(
+                CompositionPorts.Output<FlowValue>(
                     TimersCompositionPortNames.Output,
                     node.Output)
             ],
-            events: node.Events,
-            errors: node.Errors));
+            events: node.Events));
+    }
+
+    private static ValueTask<ComposedNode> CreateFlowValueTimerDelayNode(
+        CompositionNodeFactoryContext context)
+    {
+        var settings = context.BindConfiguration<TimerDelaySettings>();
+        var clock = context.GetResource<TimeProvider>(
+            TimersCompositionResourceNames.Clock);
+        var node = new FlowValueTimerDelayNode(settings, clock);
+
+        return ValueTask.FromResult(ComposedNode.Create(
+            node,
+            inputs:
+            [
+                CompositionPorts.Input<FlowValue>(
+                    TimersCompositionPortNames.Input,
+                    node.Input)
+            ],
+            outputs:
+            [
+                CompositionPorts.Output<FlowResult<FlowValue>>(
+                    TimersCompositionPortNames.Output,
+                    node.Output)
+            ],
+            events: node.Events));
+    }
+
+    private static ValueTask<ComposedNode> CreateFlowValueTimerThrottleNode(
+        CompositionNodeFactoryContext context)
+    {
+        var settings = context.BindConfiguration<TimerThrottleSettings>();
+        var clock = context.GetResource<TimeProvider>(
+            TimersCompositionResourceNames.Clock);
+        var node = new FlowValueTimerThrottleNode(settings, clock);
+
+        return ValueTask.FromResult(ComposedNode.Create(
+            node,
+            inputs:
+            [
+                CompositionPorts.Input<FlowValue>(
+                    TimersCompositionPortNames.Input,
+                    node.Input)
+            ],
+            outputs:
+            [
+                CompositionPorts.Output<FlowResult<FlowValue>>(
+                    TimersCompositionPortNames.Output,
+                    node.Output)
+            ],
+            events: node.Events));
+    }
+
+    private static ValueTask<ComposedNode> CreateFlowValueTimerDebounceNode(
+        CompositionNodeFactoryContext context)
+    {
+        var settings = context.BindConfiguration<TimerDebounceSettings>();
+        var clock = context.GetResource<TimeProvider>(
+            TimersCompositionResourceNames.Clock);
+        var node = new FlowValueTimerDebounceNode(settings, clock);
+
+        return ValueTask.FromResult(ComposedNode.Create(
+            node,
+            inputs:
+            [
+                CompositionPorts.Input<FlowValue>(
+                    TimersCompositionPortNames.Input,
+                    node.Input)
+            ],
+            outputs:
+            [
+                CompositionPorts.Output<FlowResult<FlowValue>>(
+                    TimersCompositionPortNames.Output,
+                    node.Output)
+            ],
+            events: node.Events));
     }
 
     private static ValueTask<ComposedNode> CreateTimerDelayNode<TInput>(