+++ src/FluxFlow.Components.Timers/TimerErrorCodeNames.cs
@@ -1,0 +1,9 @@
+namespace FluxFlow.Components.Timers;
+
+public static class TimerErrorCodeNames
+{
+    public const string MissingInput = "timer.missing_input";
+    public const string DebounceFailed = "timer.debounce_failed";
+    public const string DelayFailed = "timer.delay_failed";
+    public const string ThrottleFailed = "timer.throttle_failed";
+}